refactor: restrcuture package so that we control what we export
