+++ lib/src/models/flutter_datawedge_exception.dart
