+++ lib/src/models/scan_result.dart
