+++ lib/src/models/scan_result.freezed.dart
