+++ lib/src/models/scanner_status.dart
