+++ lib/src/models/scanner_status.freezed.dart
